Use simpler method of getting hot keys.
We could also just build up a list of keys and then sort, but this ends
up being about the same level of code and readability.  Using this
strategy also allows us to turn the heap into a priority queue later if we
decide we want to do things like limit the number of keys tracked (probably
most useful when tracking raw key hits instead of using regexps, where the
cardinality and memory footprint might be high).