Fix issue with truncated packets.
Previously, truncated packets were handled just fine, because
we didn't attempt to continue looking for commands in a packet
after we found the initial one.

In 1.0.3, we started reading the byte count for those commands
that also passed a value.  If a full value command and carriage-
return/newline combo were not found, a panic would occur and
mcsauna would crash.  This commit ensures that parsing only
continues for packets with full value commands present.